+++ src/main/kotlin/com/github/naoyukik/intellijplugingithubnotifications/domain/GitHubNotificationRepository.kt
@@ -1,7 +1,8 @@
 package com.github.naoyukik.intellijplugingithubnotifications.domain
 
 import com.github.naoyukik.intellijplugingithubnotifications.domain.model.GitHubNotification
-import com.github.naoyukik.intellijplugingithubnotifications.domain.model.NotificationDetail
+import com.github.naoyukik.intellijplugingithubnotifications.domain.model.NotificationDetailResponse
+import java.time.ZonedDateTime
 
 interface GitHubNotificationRepository {
     fun fetchNotifications(ghCliPath: String): List<GitHubNotification>
@@ -10,5 +11,14 @@ interface GitHubNotificationRepository {
         ghCliPath: String,
         repositoryName: String,
         detailApiPath: String,
-    ): NotificationDetail
+    ): NotificationDetailResponse
+    fun fetchLatestNotifications(
+        ghCliPath: String,
+        previousTime: ZonedDateTime,
+    ): List<GitHubNotification>
+    fun fetchLatestNotificationsByRepository(
+        ghCliPath: String,
+        repositoryName: String,
+        previousTime: ZonedDateTime,
+    ): List<GitHubNotification>
 }