+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ## [Unreleased]
 
+### Added
+
+#### System
+
+- Add mockk
+
 ## [0.5.0] - 2025-01-30
 
 ### Added