+++ CHANGELOG.md
@@ -10,7 +10,8 @@
   - With this update, except for the first time after startup, new notifications will appear only when there is an update in Notifications on GitHub
   - Limitations
     - Due to GitHub API specifications, you will not be notified if it has been read.
-- Add unread notifications column
+- Add support for including read notifications in fetch
+  - Add unread notifications column
 
 ### Fixed