+++ CHANGELOG.md
@@ -4,11 +4,14 @@
 
 ## [Unreleased]
 
+## [0.7.0] - 2025-02-21
+
 ### Added
 
 - Setting a limit on the number of results
 
 ### Fix
+
 - Remade the settings screen
 
 ## [0.6.0] - 2025-02-16
@@ -87,7 +90,8 @@ Initial release of the GitHub Notifications IntelliJ Plugin.
 
 Done is better than perfect.
 
-[Unreleased]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.4.3...v0.5.0
 [0.4.3]: https://github.com/naoyukik/intellij-plugin-github-notifications/compare/v0.4.2...v0.4.3